remove e2e useInterval
